Major update, fix how SDB operates
This commit touches three files.

in package.lisp: export all relevant symbols

in readme.org: Update the readme to communicate how SDB is written and the
various errata relevant to its usage. Update the example configuration and the
documentation in the readme.

in swm-debugger-mode.lisp: reorganize the file to make more sense, remove unused
functionality, break out behavior into methods specializing on a mixin, refactor
large portions of the codebase to make more sense and eliminate bugs, add ways
to specify which debugger hook to use, fix plethora of bugs including: not
invoking restarts; failing to grab keypresses; read and evaluate user input in a
more sane manner; force the usage of the interactor for reading input;